Update xml-filereader.cpp
With the compiler options "-ffast-math -fsingle-precision-constant" the Windows Mingw compiler causes an inaccuracy.

+++ src/devices/filereaders/xml-filereader/xml-filereader.cpp
@@ -90,7 +90,7 @@ XmlFileReader::XmlFileReader(const QString & iFilename)
   containerLabel->setText(theDescriptor->container);
   iqOrderLabel->setText(theDescriptor->iqOrder);
   byteOrderLabel->setText(theDescriptor->byteOrder);
-  frequencyDisplay->display(theDescriptor->blockList[0].frequency / 1000.0);
+  frequencyDisplay->display(theDescriptor->blockList[0].frequency / (double)1000.0);
   typeofUnitLabel->setText(theDescriptor->blockList[0].typeofUnit);
   modulationtypeLabel->setText(theDescriptor->blockList[0].modType);
